pattern-reliability/requirements.txt
matplotlib
numpy
opencv-python==4.6.0.66
Pillow~=8.2.0
scikit-image~=0.18.1
scikit-learn==1.1.2
torch==1.12.1
torchvision==0.13.1
tqdm~=4.59.0